lint

+++ contracts/test/TestPriceOracle.sol
@@ -4,13 +4,13 @@ pragma solidity ^0.8.0;
 import "../oracle/price/IPriceOracle.sol";
 
 contract TestPriceOracle is IPriceOracle {
-    uint private _price;
+    uint256 private _price;
 
-    function setPrice(uint price_) public {
+    function setPrice(uint256 price_) public {
         _price = price_;
     }
 
-    function price() external view override returns (uint) {
+    function price() external view override returns (uint256) {
         return _price;
     }
 }
